+++ utils/once/once_with_error.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package oncewitherror
+package once
 
 import "sync"
